requirements.txt
streamlit
py-feat==0.4.0
opencv-python-headless
pandas
numpy<2.0.0
scipy
scikit-learn
matplotlib